@@ -12,18 +12,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.9",
-    "@angular/common": "5.2.9",
-    "@angular/compiler": "5.2.9",
-    "@angular/core": "5.2.9",
-    "@angular/forms": "5.2.9",
-    "@angular/http": "5.2.9",
-    "@angular/platform-browser": "5.2.9",
-    "@angular/platform-browser-dynamic": "5.2.9",
-    "@angular/router": "5.2.9",
-    "@ngrx/store": "^5.2.0",
-    "angular2-jwt": "^0.2.2",
-    "auth0-lock": "^11.4.0",
+    "@angular/animations": "^6.0.7",
+    "@angular/common": "^6.0.7",
+    "@angular/compiler": "^6.0.7",
+    "@angular/core": "^6.0.7",
+    "@angular/forms": "^6.0.7",
+    "@angular/http": "^6.0.7",
+    "@angular/platform-browser": "^6.0.7",
+    "@angular/platform-browser-dynamic": "^6.0.7",
+    "@angular/router": "^6.0.7",
+    "@auth0/angular-jwt": "^2.0.0",
+    "@ngrx/store": "*",
+    "auth0-js": "^9.6.1",
+    "auth0-lock": "^11.7.2",
     "chart.js": "2.7.2",
     "core-js": "2.5.3",
     "file-saver": "^1.3.3",
@@ -33,36 +34,38 @@
     "jquery": "3.3.1",
     "lz-string": "^1.4.4",
     "moment": "2.21.0",
-    "ng2-ace-editor": "^0.3.1",
     "ng2-dragula": "^1.5.0",
-    "ngx-papaparse": "^2.1.0",
-    "primeng": "5.2.3",
+    "ngx-papaparse": "^2.1.4",
+    "papaparse": "4.4.0",
+    "primeicons": "1.0.0-beta.9",
+    "primeng": "6.0.0",
     "quill": "1.3.6",
-    "rxjs": "5.5.7",
+    "rxjs": "6.2.1",
     "web-animations-js": "2.3.1",
-    "zone.js": "0.8.20"
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.3",
-    "@angular/compiler-cli": "5.2.9",
-    "@types/jasmine": "2.8.6",
-    "@types/jasminewd2": "2.0.3",
-    "@types/node": "9.4.7",
+    "@angular-devkit/build-angular": "*",
+    "@angular/cli": "*",
+    "@angular/compiler-cli": "^6.0.7",
+    "@angular/language-service": "^6.0.7",
     "@types/createjs": "0.0.29",
     "@types/filesaver": "0.0.30",
-    "@types/signalr": "^2.2.32",
+    "@types/jasmine": "*",
+    "@types/jasminewd2": "*",
+    "@types/node": "8.9.5",
+    "@types/signalr": "*",
     "codelyzer": "4.2.1",
-    "jasmine-core": "3.1.0",
+    "jasmine-core": "2.99.1",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "2.0.0",
+    "karma": "1.7.1",
     "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-jasmine": "1.1.1",
-    "karma-jasmine-html-reporter": "1.0.0",
-    "karma-coverage-istanbul-reporter": "1.4.2",
+    "karma-coverage-istanbul-reporter": "2.0.1",
+    "karma-jasmine": "1.1.2",
+    "karma-jasmine-html-reporter": "0.2.2",
     "protractor": "5.3.0",
     "ts-node": "5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "typescript": "2.7.2"
   }
 }
