@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import 'rxjs/add/operator/merge';
+import { merge } from 'rxjs/operators';
 
 import { RealmHubSenderService } from '../shared/services/realm-hub-sender.service';
 import { WebApiService } from '../shared/services/rest/web-api.service';
@@ -15,10 +15,10 @@ export class ErrorDialogComponent implements OnInit {
   errorMessage: string;
 
   constructor(private realmHubSenderService: RealmHubSenderService,
-              private webApiService: WebApiService) {
+    private webApiService: WebApiService) {
     this.realmHubSenderService.errorSubject.
-      merge(this.webApiService.errorSubject).
-      subscribe(err => {
+      pipe(merge(this.webApiService.errorSubject))
+      .subscribe(err => {
         this.errorMessage = err;
         this.display = true;
       });
