@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import {Store} from '@ngrx/store';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 
 import {AppStore} from '../app-store';
 import {VisibleAreaActions} from '../actions/visible-area.actions';
