@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Subject } from 'rxjs/Subject';
+import { Subject } from 'rxjs';
 
 import { saveAs } from 'file-saver';
 import { PapaParseService, PapaParseResult } from 'ngx-papaparse';
