@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { AsyncSubject } from 'rxjs/AsyncSubject';
+import { AsyncSubject } from 'rxjs';
 
 import { IBinaryExpression, IExpression, IIdentifier, ILiteral,
          IMemberExpression, IUnaryExpression, ICallExpression, IConditionalExpression } from '../../models/jsep';
